Create boar prefab and complete character
Boar has basic UI, it goes to the player and then hits it, can be damaged and die.
The character can now punch (G key) and kick (H key). It changes the range and damage, not animated but prints the type of the attack.
It can die and it will disapear for now.